src/python/cloud_functions/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-datastore==1.8.0

